Update to build.sh to allow compilation of additional files
Update build.sh to look for ADD_FILE_PATH in the config.rb file.
ADD_FILE_PATH should be a relative path to the config.rb that points to
another SCSS directory with a config file to be built when this file is
built.

Best use case: Global partials included by multiple projects. When you
make a change to the global partial, compile all projects as well.